add possibility to update a bet

+++ WebApi/Controllers/BetController.cs
@@ -50,13 +50,31 @@ public class BetController(IBetService betService) : ControllerBase
     }
     
     
-    // TODO : Delete Before prod
+    // TODO : Set if user is admin
     [HttpPost]
     [Route("delete/{betId:guid}")]
     public async Task<ActionResult> DeleteById(Guid betId)
     {
+        if (User.GetNameIdentifierId() != "floguerin156@gmail.com")
+            return Forbid();
+
         await betService.DeleteAsync(betId);
         
         return Ok();
     }
+    
+    // TODO : Set if user is admin
+    [HttpPut]
+    [Route("update/{betId:guid}")]
+    public async Task<ActionResult> UpdateAsync(Guid betId, UpdateUserBetRequest request)
+    {
+        if (User.GetNameIdentifierId() != "floguerin156@gmail.com")
+            return Forbid();
+
+        await betService.UpdateAsync(betId, request);
+        
+        return Ok();
+    }
+    
+    
 }
